救急医療の東京ルールの適用件数; Rename components/cards/TokyoRulesApplicationNumberCard.vue to components/index/CardsMonitoring/TokyoRulesApplicationNumber/Card.vue
